Move NodeTopology into node package
