+++ src/main/java/thorpe/luke/network/simulation/worker/Worker.java
@@ -3,9 +3,9 @@ package thorpe.luke.network.simulation.worker;
 import java.nio.file.Path;
 import thorpe.luke.log.Logger;
 import thorpe.luke.network.packet.Packet;
-import thorpe.luke.network.simulation.NodeTopology;
 import thorpe.luke.network.simulation.mail.Mailbox;
 import thorpe.luke.network.simulation.mail.PostalService;
+import thorpe.luke.network.simulation.node.NodeTopology;
 import thorpe.luke.time.Clock;
 import thorpe.luke.util.ThreadNameGenerator;
 import thorpe.luke.util.error.ExceptionListener;
