+++ src/main/java/thorpe/luke/network/simulation/worker/WorkerAddressBook.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import thorpe.luke.log.Logger;
-import thorpe.luke.network.simulation.NodeTopology;
 import thorpe.luke.network.simulation.mail.Mailbox;
 import thorpe.luke.network.simulation.mail.PostalService;
+import thorpe.luke.network.simulation.node.NodeTopology;
 import thorpe.luke.time.Clock;
 import thorpe.luke.util.GarbageCollector;
 import thorpe.luke.util.Prunable;
